🚚 [Move]: #59 Result / Option を src/utils/ に移動
`src/lib/` はアプリ固有ロジック (API クライアント / 外部 SDK ラッパ等) の
置き場であり、フレームワーク非依存の汎用プリミティブである Result / Option
は別レイヤとして分離するのが妥当なため、`src/utils/result/` /
`src/utils/option/` に移動。invoke ラッパ群の import パスを `@/lib/result`
→ `@/utils/result` に追従、CLAUDE.md のディレクトリ構成にも `utils/` /
`domains/` を追記して責務を明文化した。
