+++ src/lib/tauri/createTask.ts
@@ -1,6 +1,6 @@
 import type { Priority } from "@/domains/priority";
-import type { Result } from "@/lib/result";
 import type { Task } from "@/types/task";
+import type { Result } from "@/utils/result";
 import { invokeWrapped } from "./invokeWrapped";
 import type { TauriError } from "./tauriError";
