+++ src/lib/tauri/deleteTask.ts
@@ -1,4 +1,4 @@
-import type { Result } from "@/lib/result";
+import type { Result } from "@/utils/result";
 import { invokeWrapped } from "./invokeWrapped";
 import type { TauriError } from "./tauriError";
